DS1 Merged a modified version of Kathy's deep sea mod
Note that we generate 2 "bands" of ocean rather than 1. The rest of what was previously ocean is now deep ocean.
This change is intended to allow an intermediate form of seafaring, probably enabled by either medieval ships or the compass tech.
The textures provided did not blend well so I chose to keep the "stock" ocean textures. This is not desirable since we
should be able to distinguish graphically between ocean and deep ocean